Move organization details settings to its own file
